Data-Analytics-Portfolio/requirements.txt
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.0.9
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
et_xmlfile==2.0.0
fonttools==4.59.1
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kaleido==1.0.0
kiwisolver==1.4.9
llvmlite==0.44.0
logistro==1.1.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multimethod==1.12
narwhals==2.1.2
networkx==3.5
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
pandas==2.3.1
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
plotly==6.3.0
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
statsmodels==0.14.5
sweetviz==2.3.1
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
visions==0.8.1
win32_setctime==1.2.0
wordcloud==1.9.4
ydata-profiling==4.16.1
